Extract sun session store
