+++ service-worker.js
@@ -297,6 +297,8 @@ const APP_SHELL = [
   // Sun + light modules are statically reachable from the app shell.
   '/js/sun.js',
   '/js/sun-active-session.js',
+  '/js/sun-session-model.js',
+  '/js/sun-sessions-store.js',
   '/js/sun-session-ui.js',
   '/js/sun-body-silhouette.js',
   '/js/sun-ai-analysis.js',
