+++ tests/test-correctness-phase2.js
@@ -148,6 +148,8 @@ const pwaAppShellAssets = [
   '/js/light-tools.js',
   '/js/sun.js',
   '/js/sun-active-session.js',
+  '/js/sun-session-model.js',
+  '/js/sun-sessions-store.js',
   '/js/sun-session-ui.js',
   '/js/sun-spectrum.js',
   '/js/sun-uvdata.js',
